fixed order of :: names on ccc.
added aman->RecycleArena and ?recyclearena. this required a bunch of
changes and reorganization:

	added an ARENA_CLOSING state after running that acts mostly like
	running, but doesn't allow new players in.

	the reaper moves arenas from closing to the next state just like
	running, but it allows them to move even if some players are still
	referencing the arena (but it forces the arena to be resurrected).

	added support for resurrecting arenas. this just means that when the
	arena dies, it immediately gets recreated from scratch (in the same
	memory). this fixes a long-standing issue with players who enter an
	arena that's in the final stages of closing. (previously I think it
	created a new arena with the same name; now it just marks the arena
	to be resurrected, and the player waits to enter it as usual.)

	changed oldarena field in Player to newarena and changed its
	meaning. now arena is valid in all states (even after leaving the
	arena, until S_LOGGEDIN is reached again), and newarena is used as
	the waiting to enter an arena trigger.

	common code was factored out of LeaveArena into a function that
	correctly handles players leaving an arena before they fully entered
	it, by moving their state to the correct state on the leaving-arena
	sequence.

	a new player state was added to differentate between doing leaving
	callbacks and saving arena data to the db, to help do the above
	correctly.

	decreased reap arena timer to < 2 seconds, since ?recyclearena has
	to wait for one iteration of it.


fixed initial ship bug that caused cont to crash hard.

cleaned up process_player_states to combine the inital switch
statements.

a few #include fixes and name changes to support compiling on cygwin (no
makefile yet).

changed ?listmods to ?listmod.

+++ src/main.c
@@ -1,6 +1,7 @@
 
 /* dist: public */
 
+#include <limits.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -214,7 +215,7 @@ local int do_chroot(void)
 		return -1;
 	}
 
-	printf("Changed root directory and set uid to %d\n", uid);
+	printf("Changed root directory and set uid to %d\n", (int)uid);
 	return 0;
 }
 #else
